+++ aws-datastore/src/main/java/com/amplifyframework/datastore/storage/sqlite/SQLiteStorageHelper.java
@@ -204,8 +204,8 @@ final class SQLiteStorageHelper extends SQLiteOpenHelper implements ModelUpdateS
             while (cursor.moveToNext()) {
                 tablesToDrop.add(cursor.getString(0));
             }
-
             sqliteDatabase.beginTransaction();
+            sqliteDatabase.execSQL("PRAGMA foreign_keys = OFF;");
             for (String table : tablesToDrop) {
                 // Android SQLite creates system tables to store metadata
                 // and all the system tables have the prefix SQLITE_SYSTEM_TABLE_PREFIX.
@@ -215,6 +215,7 @@ final class SQLiteStorageHelper extends SQLiteOpenHelper implements ModelUpdateS
                 sqliteDatabase.execSQL("DROP TABLE IF EXISTS " + Wrap.inBackticks(table));
                 LOG.debug("Dropped table: " + table);
             }
+            sqliteDatabase.execSQL("PRAGMA foreign_keys = ON;");
             sqliteDatabase.setTransactionSuccessful();
             sqliteDatabase.endTransaction();
         } finally {